changed variables

+++ arrays/BasicPractises/largestNumber.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main() {
    int array[] = {1, 2, 3, 10, 11, 14, 9, 16, 3};
-   int largestNumber = INT8_MIN;
+   int largestNumber = INT32_MIN;
    int arraySize = sizeof(array) / sizeof(array[0]);
 
    for (int i = 0; i < arraySize; i++) {
